+++ .travis.yml
@@ -1,6 +1,6 @@
 language: node_js
 matrix:
-  include: 
+  include:
     - os: linux
       dist: trusty
       node_js: stable
@@ -23,7 +23,7 @@ before_script:
   - export DISPLAY=:99.0
   - if [ "${TRAVIS_OS_NAME}" == "linux" ]; then ( scripts/setup_linux_env.sh  ) fi
   - if [ "${TRAVIS_OS_NAME}" == "osx" ]; then ( scripts/setup_osx_env.sh ) fi
-  - sleep 2 
+  - sleep 2
 
 script:
   - set -x
